Separate folders for legacy and current library integrations.
